refactor: single source for the community key constant
