+++ src/config.ts
@@ -1,3 +1,5 @@
+export const COMMUNITY_KEY_ID = "cea10b57de8e060ed1a180a00c2bc717a2ab4f231d88fd33ffa6a50a04f23b6e"
+
 export const config = {
 	submission: {
 		maxVariantsPerUserPerVideo: 3,
@@ -29,9 +31,7 @@ export const config = {
 	linking: {
 		stateTtlSeconds: 600, // OAuth round-trip window for the Discord link flow
 		discordScope: "identify",
-		blacklistedKeyIds: new Set<string>([
-			"cea10b57de8e060ed1a180a00c2bc717a2ab4f231d88fd33ffa6a50a04f23b6e",
-		]),
+		blacklistedKeyIds: new Set<string>([COMMUNITY_KEY_ID]),
 	},
 
 	reputation: {
